pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "marimo-tutorials"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "dspy>=3.0.1",
    "kuzu>=0.11.1",
    "marimo>=0.14.17",
    "polars>=1.32.3",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=2.2.2",
    "pydantic>=2.7.1",
    "scikit-learn>=1.3.0",
    "cachetools>=5.3.0",
    "numpy>=1.24.0",
    "pandas>=2.3.3",
    "matplotlib>=3.7.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
benchmark = [
    "tiktoken>=0.5.0",
]

[tool.setuptools]
packages = ["cache_method", "text_2_cypher"]
